move to enums folder
