+++ packages/cli/commands/customObject/create.js
@@ -8,7 +8,7 @@ const {
 const { trackCommandUsage } = require('../../lib/usageTracking');
 const { getAccountId } = require('../../lib/commonOpts');
 const { batchCreateObjects } = require('@hubspot/cli-lib/api/customObject');
-const { EXIT_CODES } = require('../../lib/exitCodes');
+const { EXIT_CODES } = require('../../lib/enums/exitCodes');
 
 exports.command = 'create <name> <definition>';
 exports.describe = 'Create custom object instances';
