+++ packages/cli/commands/customObject/schema/create.js
@@ -14,7 +14,7 @@ const {
   createSchema: createSchemaFromHubFile,
 } = require('@hubspot/cli-lib/api/fileTransport');
 const { getHubSpotWebsiteOrigin } = require('@hubspot/cli-lib/lib/urls');
-const { EXIT_CODES } = require('../../../lib/exitCodes');
+const { EXIT_CODES } = require('../../../lib/enums/exitCodes');
 
 exports.command = 'create <definition>';
 exports.describe = 'Create a custom object schema';
