+++ packages/cli/commands/fetch.js
@@ -13,7 +13,7 @@ const {
 const { resolveLocalPath } = require('../lib/filesystem');
 const { validateMode, loadAndValidateOptions } = require('../lib/validation');
 const { trackCommandUsage } = require('../lib/usageTracking');
-const { EXIT_CODES } = require('../lib/exitCodes');
+const { EXIT_CODES } = require('../lib/enums/exitCodes');
 
 exports.command = 'fetch <src> [dest]';
 exports.describe =
