+++ packages/cli/commands/marketplaceValidate/validateTheme.js
@@ -19,7 +19,7 @@ const {
 const { applyValidators } = require('../../lib/validators/applyValidators');
 const MARKETPLACE_VALIDATORS = require('../../lib/validators');
 const { VALIDATION_RESULT } = require('../../lib/validators/constants');
-const { EXIT_CODES } = require('../../lib/exitCodes');
+const { EXIT_CODES } = require('../../lib/enums/exitCodes');
 
 exports.command = 'marketplace-validate <src>';
 exports.describe = 'Validate a theme for the marketplace';
