+++ packages/cli/commands/watch.js
@@ -15,7 +15,7 @@ const {
 } = require('../lib/commonOpts');
 const { validateMode, loadAndValidateOptions } = require('../lib/validation');
 const { trackCommandUsage } = require('../lib/usageTracking');
-const { EXIT_CODES } = require('../lib/exitCodes');
+const { EXIT_CODES } = require('../lib/enums/exitCodes');
 
 exports.command = 'watch <src> <dest>';
 exports.describe =
